+++ src/models/prismaClient.js
@@ -1,0 +1,3 @@
+const { PrismaClient } = require('@prisma/client');
+const prisma = new PrismaClient();
+module.exports = prisma;
