+++ src/public/card.css
@@ -1,0 +1,254 @@
+/* General Reset */
+* {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+}
+
+body {
+  font-family: 'Arial', sans-serif;
+  text-align: center;
+  color: #333;
+  height: 100vh;
+  background: radial-gradient(ellipse at center, #000000 0%, #000000 100%);
+  overflow: hidden;
+  filter: drop-shadow(0 0 10px rgb(187, 164, 164));
+  font-size: 16px
+    /* background-image: url(./images/snowfall.gif);
+    background-repeat: no-repeat;
+    background-size: cover;
+    background-position: center; */
+}
+
+/* Header */
+header {
+  background-color: #ff6f61;
+  color: #fff;
+  padding: 1rem;
+  font-size: 2rem;
+  position: sticky;
+  font-weight: bold;
+}
+
+nav {
+  display: flex;
+  justify-content: flex-end;
+  background-color: #f95e4e;
+  padding: 0.5rem;
+}
+
+nav a {
+  color: white;
+  text-decoration: none;
+  margin-left: 1rem;
+  font-weight: bold;
+  font-size: 1.5rem;
+}
+
+section {
+  margin: 2rem auto;
+  max-width: 900px;
+}
+
+.box {
+  max-width: 1000px;
+}
+
+h1 {
+  color: #6ce40a;
+  font-family: 'Georgia', serif;
+  font-size: 1.8rem;
+  margin-bottom: 2rem;
+}
+
+
+
+.title {
+  color: #6ce40a;
+  font-family: 'Georgia', serif;
+  font-size: 1.8rem;
+  margin-bottom: 2rem;
+}
+
+article {
+  display: inline-flex;
+  flex-direction: column;
+  justify-content: center;
+  background-color: #fba494;
+  border-radius: 10px;
+  padding: 1rem;
+  margin: 1rem;
+  width: 250px;
+  height: 200px;
+  box-shadow: 0 4px 6px rgba(82, 255, 96, 0.1);
+}
+
+
+article img {
+  width: 100%;
+  border-radius: 10px;
+  margin-bottom: 0.5rem;
+}
+
+button {
+  background-color: #ff6f61;
+  color: #fff;
+  border: none;
+  padding: 0.5rem 1rem;
+  border-radius: 5px;
+  cursor: pointer;
+  font-size: 1rem;
+}
+
+button:hover {
+  background-color: #f95e4e;
+}
+
+/* Popup Modal */
+.modal {
+  display: none;
+  position: fixed;
+  z-index: 1000;
+  left: 0;
+  top: 0;
+  width: 100%;
+  height: 100%;
+  overflow: auto;
+  background-color: rgba(0, 0, 0, 0.6);
+}
+
+.modal-content {
+  background-color: #fff;
+  margin: 10% auto;
+  padding: 20px;
+  border: 9px solid #000000;
+  width: 300px;
+  border-radius: 10px;
+  text-align: center;
+}
+
+.modal-content img {
+  width: 100%;
+  border-radius: 10px;
+  margin-bottom: 1rem;
+}
+
+.modal-content textarea {
+  width: 100%;
+  height: 80px;
+  margin: 1rem 0;
+  border: 1px solid #ccc;
+  border-radius: 5px;
+  padding: 0.5rem;
+  font-size: 1rem;
+}
+
+.close {
+  color: #aaa;
+  float: right;
+  font-size: 28px;
+  font-weight: bold;
+  cursor: pointer;
+}
+
+.close:hover,
+.close:focus {
+  color: black;
+}
+
+article h2 {
+  font-size: 1.6rem;
+  margin-bottom: 0.5rem;
+  color: #fff;
+}
+
+article p {
+  font-size: 1.4rem;
+  color: #fff;
+}
+
+/* Footer & Templates Section */
+footer {
+  font-family: 'Georgia', serif;
+  color: #396218;
+  margin: 2rem auto 1rem;
+  font-size: 1.4rem;
+}
+
+section.templates {
+  margin: 2rem auto;
+  max-width: 900px;
+}
+
+section.templates article {
+  display: inline-block;
+  background-color: #fba494;
+  padding: 1rem;
+  border-radius: 10px;
+  width: 200px;
+  margin: 1rem;
+  text-align: center;
+  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+}
+
+section.templates img {
+  width: 100%;
+  border-radius: 10px;
+  margin-bottom: 0.5rem;
+}
+
+section.templates button {
+  background-color: #ff6f61;
+  color: #fff;
+  border: none;
+  padding: 0.5rem 1rem;
+  border-radius: 5px;
+  cursor: pointer;
+  font-size: 1rem;
+}
+
+section.templates button:hover {
+  background-color: #f95e4e;
+}
+
+.modal-content {
+  background-color: #3f1111;
+  color: #ffffff;
+  border-radius: 8px;
+  padding: 20px;
+
+}
+
+.modal-content h3,p{
+  margin: 1rem;
+}
+
+.social-icons img {
+  width: 30px;
+  height: 30px;
+}
+
+:root {
+  --snow-count: 20; 
+}
+
+.snow {
+  position: absolute;
+  width: 10px;
+  height: 10px;
+  background: white;
+  border-radius: 50%;
+  animation: fall linear infinite;
+}
+
+@keyframes fall {
+  0% {
+    transform: translate(var(--x-start), -10px) scale(var(--scale));
+  }
+  50% {
+    transform: translate(var(--x-mid), var(--y-mid)) scale(var(--scale));
+  }
+  100% {
+    transform: translate(var(--x-end), 100vh) scale(var(--scale));
+  }
+}
