+++ src/public/viewcard.html
@@ -1,0 +1,57 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>View Your Christmas Card</title>
+    <style>
+        body {
+            font-family: Arial, sans-serif;
+            display: flex;
+            flex-direction: column;
+            justify-content: center;
+            align-items: center;
+            height: 100vh;
+            background-color: #f5f5f5;
+        }
+        .card {
+            text-align: center;
+            background-color: white;
+            padding: 20px;
+            border-radius: 10px;
+            box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.1);
+        }
+        .card img {
+            max-width: 400px;
+            margin-bottom: 20px;
+        }
+        .message {
+            font-size: 18px;
+            font-style: italic;
+            color: #555;
+        }
+    </style>
+</head>
+<body>
+    <div class="card">
+        <img id="cardImage" src="" alt="Christmas Card">
+        <p class="message" id="userMessage"></p>
+    </div>
+
+    <script>
+        // Get the URL parameters and display the saved image and message
+        window.onload = function() {
+            const urlParams = new URLSearchParams(window.location.search);
+            const image = urlParams.get('image');
+            const message = urlParams.get('message');
+
+            if (image && message) {
+                document.getElementById('cardImage').src = decodeURIComponent(image);
+                document.getElementById('userMessage').innerText = decodeURIComponent(message);
+            } else {
+                alert("No card found!");
+            }
+        };
+    </script>
+</body>
+</html>
