+++ src/public/wall.html
@@ -1,0 +1,96 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>MemoriesWall</title>
+    <link rel="stylesheet" href="styles.css">
+    <link rel="preconnect" href="https://fonts.googleapis.com">
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+    <link href="https://fonts.googleapis.com/css2?family=Agu+Display&display=swap" rel="stylesheet">
+</head>
+<body>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <section class="header">
+        <nav>
+            <p id="treeofsharingtitle">
+                <a href="index.html">Tree Of Sharing</a>
+            </p>            
+            <div class="nav-links">
+                <ul>
+                    <li><a href="wall.html">Memories Wall</a></li>
+                    <li><a href="card.html">Send a Card</a></li>
+                    <li><a href="wishes.html">Add a Wish</a></li>
+                </ul>
+            </div>
+        </nav>
+    </section>
+
+    <div class="gallery-container">
+        <div class="gallery">
+            <img src="/images/day1.jpg" alt="Friends">
+            <div class="gdescription">Day 1</div>
+        </div>
+            
+        <div class="gallery">
+            <img src="/images/day2.jpg" alt="Latte">
+            <div class="gdescription">Day 2</div>
+        </div>
+        
+        <div class="gallery">
+            <img src="./images/day3.jpg" alt="Woman who is painting">
+            <div class="gdescription">Day 3</div>
+        </div>
+    </div>
+
+    <div class="gallery-container">
+        <div class="gallery">
+            <img src="./images/day4.jpg" alt="Cooking">
+            <div class="gdescription">Day 4</div>
+        </div>
+            
+        <div class="gallery">
+            <img src="/images/day5.jpg" alt="Bubble Tea">
+            <div class="gdescription">Day 5</div>
+        </div>
+        
+        <div class="gallery">
+            <img src="/images/day6.jpg" alt="Arcade">
+            <div class="gdescription">Day 6</div>
+        </div>
+    </div>
+
+
+    <script>
+    </script>
+</body>
+</html>
