+++ src/public/wishes.html
@@ -1,0 +1,69 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>AddAWish</title>
+    <link rel="stylesheet" href="styles.css">
+    <link rel="preconnect" href="https://fonts.googleapis.com">
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+    <link href="https://fonts.googleapis.com/css2?family=Agu+Display&display=swap" rel="stylesheet">
+</head>
+<body>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <div class="snow"></div>
+    <section class="header">
+        <nav>
+            <p id="treeofsharingtitle">
+                <a href="index.html">Tree Of Sharing</a>
+            </p>            
+            <div class="nav-links">
+                <ul>
+                    <li><a href="wall.html">Memories Wall</a></li>
+                    <li><a href="card.html">Send a Card</a></li>
+                    <li><a href="wishes.html">Add a Wish</a></li>
+                </ul>
+            </div>
+        </nav>
+    </section>
+
+<div class="wish">
+        <p id="wishtext">Share your wish for the community and we'll display it!</p>
+        <textarea name="blog" rows="5" cols="15">Type your wishes here...</textarea>
+        <br>
+        <input type="submit" value="Submit">
+</div>
+    
+
+    <script>
+    </script>
+</body>
+</html>
